Draft version 0.1
This branch and commit shall serve as my baseline for the first draft
of my Special Problem. Sections not previously present on the previous
commit but is now available are user testing, evaluation technique,
a new main section for Expected outcome, and a new figure for a simple
prototype.

Updates on Literature review has been added as well.